# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "x402-movies",
    "version": "0.1.0",
    "description": "Movie catalog concierge over TMDB — search, details, and recommendations per query\n\nEvery paid route returns the purchased artifact in the 200 response body. Pay in USDC on Base or Solana — your client picks the rail.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "contact": {
      "name": "nirholas",
      "email": "nichxbt@gmail.com",
      "url": "https://github.com/nirholas/x402-movies"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4024",
      "description": "Local development"
    },
    {
      "url": "{baseUrl}",
      "description": "Deployed instance",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:4024"
        }
      }
    }
  ],
...
  "paths": {
    "/search": {
      "get": {
        "operationId": "searchMovies",
        "summary": "Search the movie catalogue by title, optionally narrowed to a release year ($0.001)",
        "description": "Search the movie catalogue by title, optionally narrowed to a release year\n\nPaid route — $0.001 in USDC. Pay on Base (EVM) or Solana; the 402 challenge lists both rails.",
        "tags": [
          "paid"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Title to search for, 1-200 characters.",
            "schema": {
              "type": "string"
            },
            "example": "blade runner"
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "description": "Narrow to a primary release year, 1870…2100.",
            "schema": {
              "type": "integer"
            },
            "example": 1982
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Result page, 1…500. Default 1.",
            "schema": {
              "type": "integer"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Matching titles with year, rating, vote count, overview and artwork URLs",
            "headers": {
              "X-PAYMENT-RESPONSE": {
                "description": "Base64 JSON settlement receipt.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "source",
                    "query",
                    "page",
                    "totalResults",
                    "count",
                    "results",
                    "retrievedAt"
                  ],
                  "properties": {
                    "source": {
                      "type": "string",
                      "enum": [
                        "tmdb",
                        "fixture"
                      ]
                    },
                    "query": {
                      "type": "object",
                      "properties": {
                        "q": {
                          "type": "string"
                        },
                        "year": {
                          "type": [
                            "integer",
                            "null"
                          ]
                        },
                        "page": {
                          "type": "integer"
                        }
                      }
                    },
                    "page": {
                      "type": "integer"
                    },
                    "totalResults": {
                      "type": "integer"
                    },
                    "totalPages": {
                      "type": "integer"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "retrievedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MovieSummary"
                      }
                    }
                  }
                },
                "example": {
                  "source": "fixture",
                  "query": {
                    "q": "blade runner",
                    "year": null,
                    "page": 1
                  },
                  "page": 1,
                  "totalResults": 8,
                  "totalPages": 2,
                  "count": 2,
                  "results": [
                    {
                      "id": 524916,
                      "title": "A Quiet Meridian",
                      "originalTitle": "A Quiet Meridian",
                      "overview": "A mystery about meridians, set in 2006. Deterministic fixture record — set TMDB_API_KEY for the real catalogue.",
                      "releaseDate": "2006-10-13",
                      "year": 2006,
                      "voteAverage": 5.8,
                      "voteCount": 3652,
                      "popularity": 41.907,
                      "posterUrl": null,
                      "backdropUrl": null,
                      "originalLanguage": "en",
                      "adult": false,
                      "tagline": "A Quiet meridian you will ever see.",
                      "runtimeMinutes": 98,
                      "status": "Released",
                      "budget": 97000000,
                      "revenue": 379255811,
                      "genres": [
                        "Mystery",
                        "Thriller"
                      ],
                      "productionCompanies": [
                        "Foldout Media"
                      ],
                      "spokenLanguages": [
                        "English"
                      ],
                      "homepage": null,
                      "imdbId": null,
                      "cast": [
                        {
                          "id": 6149160,
                          "name": "Iris Chen",
                          "character": "Ellis",
                          "order": 0,
                          "profileUrl": null
                        },
                        {
                          "id": 6149161,
                          "name": "Priya Raghunathan",
                          "character": "Nadia",
                          "order": 1,
                          "profileUrl": null
                        },
                        {
                          "id": 6149162,
                          "name": "Tomas Lind",
                          "character": "Rowan",
                          "order": 2,
                          "profileUrl": null
                        },
                        {
                          "id": 6149163,
                          "name": "Sofia Marchetti",
                          "character": "Sam",
                          "order": 3,
                          "profileUrl": null
                        }
                      ],
                      "crew": [
                        {
                          "id": 1324916,
                          "name": "Hanna Bergström",
                          "job": "Director",
                          "department": "Directing"
                        },
                        {
                          "id": 1334916,
                          "name": "Leo Vasquez",
                          "job": "Screenplay",
                          "department": "Writing"
                        }
                      ],
                      "watchProviders": {
                        "US": [
                          {
                            "provider": "Demo Streaming",
                            "type": "flatrate",
                            "logoUrl": null
                          },
                          {
                            "provider": "Demo Rentals",
                            "type": "rent",
                            "logoUrl": null
                          }
                        ]
                      },
                      "releaseDates": [
                        {
                          "countryCode": "US",
                          "date": "2006-10-13T00:00:00.000Z",
                          "certification": "PG-13",
                          "type": "theatrical"
                        },
                        {
                          "countryCode": "US",
                          "date": "2006-12-01T00:00:00.000Z",
                          "certification": null,
                          "type": "digital"
                        }
                      ]
                    },
                    {
                      "id": 515685,
                      "title": "After the Winter",
                      "originalTitle": "After the Winter",
                      "overview": "A adventure about winters, set in 2015. Deterministic fixture record — set TMDB_API_KEY for the real catalogue.",
                      "releaseDate": "2015-10-22",
                      "year": 2015,
                      "voteAverage": 6.7,
                      "voteCount": 1440,
                      "popularity": 89.65,
                      "posterUrl": null,
                      "backdropUrl": null,
                      "originalLanguage": "en",
                      "adult": false,
                      "tagline": "After the winter you will ever see.",
                      "runtimeMinutes": 125,
                      "status": "Released",
                      "budget": 94000000,
                      "revenue": 162875636,
                      "genres": [
                        "Adventure"
                      ],
                      "productionCompanies": [
                        "Blue Harbour Films"
                      ],
                      "spokenLanguages": [
                        "English"
                      ],
                      "homepage": null,
                      "imdbId": null,
                      "cast": [
                        {
                          "id": 6056850,
                          "name": "Yuki Tanaka",
                          "character": "Ellis",
                          "order": 0,
                          "profileUrl": null
                        },
                        {
                          "id": 6056851,
                          "name": "Ana Ferreira",
                          "character": "Nadia",
                          "order": 1,
                          "profileUrl": null
                        },
                        {
                          "id": 6056852,
                          "name": "Leo Vasquez",
                          "character": "Rowan",
                          "order": 2,
                          "profileUrl": null
                        },
                        {
                          "id": 6056853,
                          "name": "Daniel Okafor",
                          "character": "Sam",
                          "order": 3,
                          "profileUrl": null
                        }
                      ],
                      "crew": [
                        {
                          "id": 1315685,
                          "name": "Iris Chen",
                          "job": "Director",
                          "department": "Directing"
                        },
                        {
                          "id": 1325685,
                          "name": "Ana Ferreira",
                          "job": "Screenplay",
                          "department": "Writing"
                        }
                      ],
                      "watchProviders": {
                        "US": [
                          {
                            "provider": "Demo Streaming",
                            "type": "flatrate",
                            "logoUrl": null
                          },
                          {
                            "provider": "Demo Rentals",
                            "type": "rent",
                            "logoUrl": null
                          }
                        ]
                      },
                      "releaseDates": [
                        {
                          "countryCode": "US",
                          "date": "2015-10-22T00:00:00.000Z",
                          "certification": "PG-13",
                          "type": "theatrical"
                        },
                        {
                          "countryCode": "US",
                          "date": "2015-12-01T00:00:00.000Z",
                          "certification": null,
                          "type": "digital"
                        }
                      ]
                    }
                  ],
                  "retrievedAt": "2026-08-07T03:02:26.030Z"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "402": {
            "description": "Payment required (or payment rejected). The body lists both accepted rails.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRequired"
                }
              }
            }
          },
          "502": {
            "description": "Upstream data source failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/movie/{id}": {
      "get": {
        "operationId": "movieDetail",
        "summary": "Full detail bundle for one movie ($0.001)",
        "description": "Full detail bundle for one movie\n\nPaid route — $0.001 in USDC. Pay on Base (EVM) or Solana; the 402 challenge lists both rails.",
        "tags": [
          "paid"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Movie id as returned by `/search`.",
            "schema": {
              "type": "integer"
            },
            "example": 524916
          }
        ],
        "responses": {
          "200": {
            "description": "Metadata, tagline, runtime, budget/revenue, genres, cast and key crew, streaming and rental providers by country, and release dates with certifications",
            "headers": {
              "X-PAYMENT-RESPONSE": {
                "description": "Base64 JSON settlement receipt.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "source",
                    "movie",
                    "retrievedAt"
                  ],
                  "properties": {
                    "source": {
                      "type": "string",
                      "enum": [
                        "tmdb",
                        "fixture"
                      ]
                    },
                    "retrievedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "movie": {
                      "$ref": "#/components/schemas/MovieDetail"
                    }
                  }
                },
                "example": {
                  "source": "fixture",
                  "movie": {
                    "id": 524916,
                    "title": "A Quiet Meridian",
                    "originalTitle": "A Quiet Meridian",
                    "overview": "A mystery about meridians, set in 2006. Deterministic fixture record — set TMDB_API_KEY for the real catalogue.",
                    "releaseDate": "2006-10-13",
                    "year": 2006,
                    "voteAverage": 5.8,
                    "voteCount": 3652,
                    "popularity": 41.907,
                    "posterUrl": null,
                    "backdropUrl": null,
                    "originalLanguage": "en",
                    "adult": false,
                    "tagline": "A Quiet meridian you will ever see.",
                    "runtimeMinutes": 98,
                    "status": "Released",
                    "budget": 97000000,
                    "revenue": 379255811,
                    "genres": [
                      "Mystery",
                      "Thriller"
                    ],
                    "productionCompanies": [
                      "Foldout Media"
                    ],
                    "spokenLanguages": [
                      "English"
                    ],
                    "homepage": null,
                    "imdbId": null,
                    "cast": [
                      {
                        "id": 6149160,
                        "name": "Iris Chen",
                        "character": "Ellis",
                        "order": 0,
                        "profileUrl": null
                      },
                      {
                        "id": 6149161,
                        "name": "Priya Raghunathan",
                        "character": "Nadia",
                        "order": 1,
                        "profileUrl": null
                      },
                      {
                        "id": 6149162,
                        "name": "Tomas Lind",
                        "character": "Rowan",
                        "order": 2,
                        "profileUrl": null
                      },
                      {
                        "id": 6149163,
                        "name": "Sofia Marchetti",
                        "character": "Sam",
                        "order": 3,
                        "profileUrl": null
                      }
                    ],
                    "crew": [
                      {
                        "id": 1324916,
                        "name": "Hanna Bergström",
                        "job": "Director",
                        "department": "Directing"
                      },
                      {
                        "id": 1334916,
                        "name": "Leo Vasquez",
                        "job": "Screenplay",
                        "department": "Writing"
                      }
                    ],
                    "watchProviders": {
                      "US": [
                        {
                          "provider": "Demo Streaming",
                          "type": "flatrate",
                          "logoUrl": null
                        },
                        {
                          "provider": "Demo Rentals",
                          "type": "rent",
                          "logoUrl": null
                        }
                      ]
                    },
                    "releaseDates": [
                      {
                        "countryCode": "US",
                        "date": "2006-10-13T00:00:00.000Z",
                        "certification": "PG-13",
                        "type": "theatrical"
                      },
                      {
                        "countryCode": "US",
                        "date": "2006-12-01T00:00:00.000Z",
                        "certification": null,
                        "type": "digital"
                      }
                    ]
                  },
                  "retrievedAt": "2026-08-07T03:02:26.031Z"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "402": {
            "description": "Payment required (or payment rejected). The body lists both accepted rails.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRequired"
                }
              }
            }
          },
          "502": {
            "description": "Upstream data source failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "MovieSummary": {
        "type": "object",
        "required": [
          "id",
          "title",
          "overview"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Pass to `/movie/:id` and `/recommendations/:id`."
          },
          "title": {
            "type": "string"
          },
          "originalTitle": {
            "type": [
              "string",
              "null"
            ]
          },
          "overview": {
            "type": "string"
          },
          "releaseDate": {
            "type": [
              "string",
              "null"
            ],
            "description": "`YYYY-MM-DD`."
          },
          "year": {
            "type": [
              "integer",
              "null"
            ]
          },
          "voteAverage": {
            "type": [
              "number",
              "null"
            ],
            "description": "0-10."
          },
          "voteCount": {
            "type": [
              "integer",
              "null"
            ]
          },
          "popularity": {
            "type": [
              "number",
              "null"
            ]
          },
          "posterUrl": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "backdropUrl": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "originalLanguage": {
            "type": [
              "string",
              "null"
            ],
            "description": "ISO 639-1."
          },
          "adult": {
            "type": "boolean"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Machine-readable error code."
          },
          "message": {
            "type": "string",
            "description": "Human-readable description."
          }
        }
      },
      "PaymentRequirements": {
        "type": "object",
        "description": "One accepted payment rail, per the x402 specification.",
        "required": [
          "scheme",
          "network",
          "maxAmountRequired",
          "resource",
          "payTo",
          "asset"
        ],
        "properties": {
          "scheme": {
            "type": "string",
            "enum": [
              "exact"
            ]
          },
          "network": {
            "type": "string",
            "enum": [
              "base-sepolia",
              "base",
              "solana",
              "solana-devnet"
            ],
            "description": "EVM or Solana network this rail settles on."
          },
          "maxAmountRequired": {
            "type": "string",
            "description": "Price in USDC base units (6 decimals)."
          },
          "resource": {
            "type": "string",
            "format": "uri"
          },
          "description": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "payTo": {
            "type": "string",
            "description": "Receive address (0x… for EVM, base58 for Solana)."
          },
          "maxTimeoutSeconds": {
            "type": "integer"
          },
          "asset": {
            "type": "string",
            "description": "USDC contract address or SPL mint."
          },
          "outputSchema": {
            "type": "object"
          },
          "extra": {
            "type": "object",
            "description": "EVM: EIP-712 domain (`name`, `version`). Solana: `feePayer`, the sponsor that pays the SOL network fee."
          }
        }
      },
      "PaymentRequired": {
        "type": "object",
        "required": [
          "x402Version",
          "accepts"
        ],
        "properties": {
          "x402Version": {
            "type": "integer",
            "enum": [
              1
            ]
          },
          "error": {
            "type": "string"
          },
          "hint": {
            "type": "string"
          },
          "accepts": {
            "type": "array",
            "description": "Both rails: USDC on Base and USDC on Solana.",
            "items": {
              "$ref": "#/components/schemas/PaymentRequirements"
            }
          }
        }
      }
    }
  }
}